+++ R/plot_celltype_associations.r
@@ -8,9 +8,9 @@
 #'  \link[MAGMA.Celltyping]{calculate_celltype_associations} or
 #'  \link[MAGMA.Celltyping]{calculate_conditional_celltype_associations}.
 #' @param useSignificanceLine TRUE or FALSE. Should their be a vertical
-#'  line marking bonferroni signifiance?
+#'  line marking Bonferroni significance?
 #' @param savePDF TRUE or FALSE. Save figure to file or print to screen?
-#' @param fileTag String apprended to the names of the saved PDFs,
+#' @param fileTag String appended to the names of the saved PDFs,
 #'  i.e. the name of the celltype data file used
 #' @param plotDendro Should the dendrogram of celltypes be
 #'  shown alongside the figure? TRUE or FALSE.
@@ -18,20 +18,24 @@
 #' @param plotLegend Should the figure legend be displayed?
 #' @param figsDir Directory where figures should be created.
 #' @param show_plot Print to the plot(s). 
-#' @inheritParams calculate_celltype_associations
-#'
+#' @param bind_plots Bind the list of plots together using 
+#' \link[patchwork]{wrap_plots}.
+#' @param ncol Number of columns to put plots into during \code{bind_plots}.
+#' @param verbose Print messages.
+#' @inheritParams calculate_celltype_associations 
 #' @return A list of ggplot objects.
 #'
+#' @export
+#' @importFrom methods show
+#' @importFrom stats setNames
+#' @importFrom EWCE fix_celltype_names
 #' @examples
-#' res <- MAGMA.Celltyping::enrichment_results
+#' res <- enrichment_results
 #' ctAssocs <- res$`finn-a-AD.tsv.gz.35UP.10DOWN`$ctAssocsLinear
 #' ctd <- ewceData::ctd()
-#' figs <- MAGMA.Celltyping::plot_celltype_associations(
+#' figs <- plot_celltype_associations(
 #'     ctAssocs = ctAssocs,
 #'     ctd = ctd)
-#' @export
-#' @importFrom grDevices dev.off pdf 
-#' @importFrom gridExtra grid.arrange
 plot_celltype_associations <- function(ctAssocs,
                                        ctd, 
                                        useSignificanceLine = TRUE,
@@ -40,11 +44,39 @@ plot_celltype_associations <- function(ctAssocs,
                                        plotDendro = TRUE, 
                                        gwas_title = "", 
                                        plotLegend = TRUE, 
-                                       figsDir = NA,
-                                       show_plot = TRUE) {
-
-    if(is.null(ctAssocs)){
-        stop("ctAssocs is NULL.")
+                                       figsDir = NULL,
+                                       show_plot = TRUE,
+                                       bind_plots = FALSE,
+                                       ncol = 1,
+                                       verbose = TRUE) {
+    # templateR:::args2vars(plot_celltype_associations)
+    requireNamespace("ggplot2")
+    requireNamespace("patchwork")
+    
+    if(is.null(ctAssocs)) stopper("ctAssocs is NULL.")
+    #### Function to create PDF path ####
+    pdf_path <- function(figsDir,
+                         magmaPaths,
+                         annotLevel,
+                         fileTag,
+                         analysisType,
+                         type="Baseline"){
+        file.path(figsDir, 
+                   paste0(magmaPaths$gwasFileName,".",
+                          ctAssocs$upstream_kb,"UP.",
+                          ctAssocs$downstream_kb,"DOWN.",
+                          "annotLevel",annotLevel,".",
+                          type,".",
+                          fileTag,".",
+                          analysisType,'.',
+                          "pdf"
+                   )
+        ) 
+    }
+    calc_height <- function(width,
+                            ctAssocs,
+                            annotLevel){ 
+        1 + 2 * (dim(ctAssocs[[annotLevel]]$results)[1] / width)
     }
     #### CHECK: THAT RESULTS FOR ONLY ONE GWAS WERE PROVIDED  ####
     # (for more than one use magma_tileplot.r)
@@ -60,25 +92,16 @@ plot_celltype_associations <- function(ctAssocs,
         gwas_sumstats_path = ctAssocs$gwas_sumstats_path, 
         upstream_kb = ctAssocs$upstream_kb, 
         downstream_kb = ctAssocs$downstream_kb)
-    if (is.na(figsDir)) {
+    if (is.null(figsDir) || 
+        is.na(figsDir)) {
         figsDir <- magmaPaths$figs
     }
     if (!file.exists(figsDir)) {
-        dir.create(figsDir)
+        dir.create(figsDir, showWarnings = FALSE, recursive = TRUE)
     }
     #### CHECK: WAS A TITLE PROVIDED FOR THE PLOT? ####
-    if (gwas_title == "") {
-        gwas_title <- whichGWAS
-    }
-
-    # CHECK: THAT A MINIMAL SET OF COLUMN HEADERS 
-    # ARE INCLUDED IN THE RESULTS TABLE
-    requiredHeaders <- c("Celltype", "P", "log10p", "Method",
-                         "EnrichmentMode", "CONTROL", "CONTROL_label")
-
-    # Is the analysis top10%, linear or merged?
-    # print(ctAssocs[[1]]$results)
-    # print(unique(ctAssocs[[1]]$results$EnrichmentMode))
+    if (gwas_title == "") gwas_title <- whichGWAS
+    #### Is the analysis top10%, linear or merged? ####
     if (length(unique(ctAssocs[[1]]$results$EnrichmentMode)) == 1) {
         if (unique(ctAssocs[[1]]$results$EnrichmentMode) == "Linear") {
             analysisType <- "Linear"
@@ -87,110 +110,112 @@ plot_celltype_associations <- function(ctAssocs,
         }
     } else {
         analysisType <- "Merged"
+    } 
+    #### Generate the plots (for each annotation level separately) ####
+    ggplot2::theme_set(ggplot2::theme_bw()) 
+    #### Handle both new and old enrichment result formats ####
+    lvls <- seq_len(length(grep("^level.", names(ctAssocs))))
+    if(length(lvls)==0){
+        messager("No `level#` detected in results. Defaulting to level=1",
+                 v=verbose)
+        lvls <- 1
     }
-
-    # Generate the plots (for each annotation level seperately)
-    ggplot2::theme_set(ggplot2::theme_bw())
-    figures <- list()
-    for (annotLevel in seq_len(length(grep("^level.", names(ctAssocs))))) {
+    #### Iterate over results from each CTD level ####
+    figures <- lapply(stats::setNames(lvls,
+                                      paste0("level",lvls)),
+                      function(annotLevel){ 
+        messager(paste0("Plotting level ",annotLevel,":"),v=verbose)
+        messager("+ Preparing data.",v=verbose)
         # SET: NEW COLUMN COMBINING METHODS or ENRICHMENT TYPES
         ctAssocs[[annotLevel]]$results$FullMethod <- 
-            sprintf("%s %s", 
-                    ctAssocs[[annotLevel]]$results$Method,
-                    ctAssocs[[annotLevel]]$results$EnrichmentMode)
-
-        if (plotDendro == TRUE) {
-            # Order cells by dendrogram
-            ctdDendro <- get_ctd_dendro(ctd, annotLevel = annotLevel)
+            paste(ctAssocs[[annotLevel]]$results$Method,
+                  ctAssocs[[annotLevel]]$results$EnrichmentMode)
+        if (isTRUE(plotDendro)) {
+            #### Order cells by dendrogram ####
+            ctdDendro <- get_ctd_dendro(ctd = ctd, 
+                                        annotLevel = annotLevel,
+                                        expand = NULL, 
+                                        verbose = verbose)
             ctAssocs[[annotLevel]]$results$Celltype <-
-                factor(ctAssocs[[annotLevel]]$results$Celltype,
-                       levels = gsub(" |\\(|\\)", "\\.", 
-                                     ctdDendro$ordered_cells))
+                factor(EWCE::fix_celltype_names(
+                    celltypes = ctAssocs[[annotLevel]]$results$Celltype),
+                       levels = EWCE::fix_celltype_names(
+                           celltypes = ctdDendro$ordered_cells), 
+                       ordered = TRUE)
         }
-
-        a2 <- ggplot2::ggplot(ctAssocs[[annotLevel]]$results, 
-                     ggplot2::aes_string(x = "factor(Celltype)", 
-                                y = "-log10p",
-                                fill = "FullMethod")) +
-            # ggplot2::scale_y_reverse() +
-            ggplot2::geom_bar(stat = "identity", position = "dodge") +
+        #### Make the figure ####
+        messager("+ Generating figure.",v=verbose)
+        fig <- ggplot2::ggplot(ctAssocs[[annotLevel]]$results, 
+                              ggplot2::aes_string(x = "factor(Celltype)", 
+                                                  y = "-log10p",
+                                                  fill = "FullMethod")) + 
+            ggplot2::geom_bar(stat = "identity", 
+                              position = "dodge") +
             ggplot2::coord_flip() +
-            ggplot2::ylab(expression("-log"[10] * "(pvalue)")) +
-            ggplot2::xlab("")
-        a2 <- a2 +
-            ggplot2::theme(legend.position = c(0.5, 0.8)) + 
-            ggplot2::ggtitle(gwas_title) + 
-            ggplot2::theme(legend.title = ggplot2::element_blank())
-        if (plotLegend == FALSE) {
-            a2 <- a2 + ggplot2::theme(legend.position = "none")
+            ggplot2::labs(x = NULL, 
+                          y = expression("-log"[10] * "(pvalue)"),
+                          title = gwas_title, 
+                          legend = NULL) +
+            ggplot2::theme(legend.position="top") 
+        if (isFALSE(plotLegend)) {
+            fig <- fig + ggplot2::theme(legend.position = "none", 
+                                        legend.justification = c(0,0))
         }
-        if (useSignificanceLine) {
-            a2 <- a2 + 
+        if (isTRUE(useSignificanceLine)) {
+            fig <- fig + 
                 ggplot2::geom_hline(
+                    linetype="dotted",
                     yintercept = -log(
                         as.numeric(0.05 / 
                                        ctAssocs$total_baseline_tests_performed),
-                        10), colour = "black")
-        }
-        theFig <- a2 + ggplot2::theme_bw()
-
-        # If the results come from a BASELINE analysis...
+                        10), 
+                    colour = "black", 
+                    alpha=.8)
+        } 
+        #### Conditionnally format the plot ####
+        #### If the results come from a BASELINE analysis ####
         if (length(unique(ctAssocs[[1]]$results$CONTROL)) == 1) {
-            if (plotDendro == TRUE) {
-                theFig <- gridExtra::grid.arrange(
-                    a2, ctdDendro$dendroPlot,
-                   ncol = 2, 
-                   widths = c(0.8, 0.2))
-            }
-
-            if (savePDF) {
-                fName <- sprintf(
-                    "%s/%s.%sUP.%sDOWN.annotLevel%s.Baseline.%s.%s.pdf", 
-                    figsDir,
-                    magmaPaths$gwasFileName,
-                    ctAssocs$upstream_kb, 
-                    ctAssocs$downstream_kb, 
-                    annotLevel, 
-                    fileTag, 
-                    analysisType) 
-                # print("here")
-                height <-  1 + 2 * (dim(ctAssocs[[annotLevel]]$results)[1] / 10)
-                grDevices::pdf(file = fName, 
-                               width = 10, 
-                               height = height)
-                print(gridExtra::grid.arrange(
-                    a2 + ggplot2::theme_bw(), 
-                    ctdDendro$dendroPlot, 
-                    ncol = 2, widths = c(0.8, 0.2)))
-                grDevices::dev.off()
-            } else {
-                if(show_plot) print(theFig)
-            }
-            # IF THE RESULTS COME FROM A CONDITIONAL ANALYSIS
+            if (isTRUE(plotDendro)) {
+                fig <- patchwork::wrap_plots(fig,  ctdDendro$dendroPlot) 
+            } 
+            width <- 10
+            type <- "Baseline"
+        
+        #### IF THE RESULTS COME FROM A CONDITIONAL ANALYSIS ####
         } else {
-            theFig <- theFig + ggplot2::facet_wrap(~CONTROL_label)
-
-            if (savePDF) {
-                fName <- sprintf(
-                    "%s/%s.%sUP.%sDOWN.annotLevel%s.ConditionalFacets.%s.%s.pdf",
-                    figsDir,
-                    magmaPaths$gwasFileName, 
-                    ctAssocs$upstream_kb, 
-                    ctAssocs$downstream_kb, 
-                    annotLevel, 
-                    fileTag, 
-                    analysisType)
-                height <- 1 + 2 * (dim(ctAssocs[[annotLevel]]$results)[1] / 30)
-                grDevices::pdf(file = fName, 
-                               width = 25,
-                               height = height)
-                print(theFig + ggplot2::theme_bw())
-                grDevices::dev.off()
-            } else {
-                if(show_plot) print(theFig)
-            }
+            fig <- fig + ggplot2::facet_wrap(~CONTROL_label) 
+            type <- "ConditionalFacets"
+            width <- 30
         }
-        figures[[length(figures) + 1]] <- theFig
+        #### Save plot ####
+        if (isTRUE(savePDF)) { 
+            fName <- pdf_path(figsDir = figsDir, 
+                              magmaPaths = magmaPaths, 
+                              annotLevel = annotLevel,
+                              fileTag = fileTag, 
+                              analysisType = analysisType, 
+                              type = "ConditionalFacets")  
+            messager("+ Saving figure ==>",fName,v=verbose)
+            height <- calc_height(width = width,
+                                  ctAssocs = ctAssocs, 
+                                  annotLevel = annotLevel)
+            ggplot2::ggsave(filename = fName,
+                            plot = fig,
+                            width = width, 
+                            height = height)
+        } 
+        #### Show plot ####
+        if(isTRUE(show_plot)) methods::show(fig)
+        return(fig)
+    }) 
+    #### Return ####
+    if(isTRUE(bind_plots)){
+        pw <- 
+            patchwork::wrap_plots(figures, 
+                                  ncol = ncol) +
+            patchwork::plot_annotation(tag_levels = list(names(figures)))
+        return(pw)
+    } else {
+        return(figures)
     }
-    return(figures)
 }